Update removeClass with selector to find the previous row
The previous tr element that was selected is not a direct
sibling of the current row, so the removeClass call was not clearing
the previously clicked ticket. This resulted in multiple tickets
showing the shading for 'in-preview'. Update the selector to backtrack
and then find the correct list of tr elements.