Add callback to fix ticket action paths using scroll history
If history is set to scrolling the paths assume you are on
Ticket/Display.html and uses relative paths. For example AttachmentPath
is set to Attachment instead of the full path. This means the paths do
not work for the search preview and so the paths need to be set to full
paths so they work.