Modify modal offset for RT search filter in side-by-side mode
The way that the ticket list is wrapped when in preview side by side
mode causes a display issue when calculating the max-height of the
search filter modal.

Add JS to overwrite the search filter on click handler to modify the
max-height calculation so the modal displays correctly.

See commit 9263f906ea for where the height offset was introduced.